requirements.txt
faiss-cpu
langchain
langchain-community
python-dotenv
streamlit